+++ apps/api/src/services/cryptocurrency/cryptocurrency.module.ts
@@ -1,0 +1,9 @@
+import { Module } from '@nestjs/common';
+
+import { CryptocurrencyService } from './cryptocurrency.service';
+
+@Module({
+  providers: [CryptocurrencyService],
+  exports: [CryptocurrencyService]
+})
+export class CryptocurrencyModule {}